replaced bl from service to handlers, except shared bl, deleted empty service, fixed swagger xml docs
