+++ SimpleAccountService/Features/Accounts/AccountsController.cs
@@ -1,6 +1,7 @@
 ﻿using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Simple_Account_Service.Application.Models;
 using Simple_Account_Service.Features.Accounts.Commands.CreateAccount;
 using Simple_Account_Service.Features.Accounts.Commands.DeleteAccount;
 using Simple_Account_Service.Features.Accounts.Commands.UpdateAccount;
@@ -22,62 +23,30 @@ public class AccountsController(IMediator mediator) : ControllerBase
     /// <summary>
     /// Создать новый счет.
     /// </summary>
-    /// <param name="createAccountDto">Данные для создания счета:
-    /// <list type="bullet">
-    /// <item><description>OwnerId — уникальный идентификатор владельца счета (Guid).</description></item>
-    /// <item><description>Type — тип счета: 0 для Checking, 1 для Deposit, 2 для Credit.</description></item>
-    /// <item><description>Currency — валюта счета в формате ISO 4217, например "USD", "EUR".</description></item>
-    /// <item><description>InterestRate — процентная ставка (опционально).</description></item>
-    /// </list>
-    /// </param>
-    /// <returns>Возвращает созданный счет с кодом 201 Created.</returns>
-    /// <remarks>
-    /// Пример запроса:
-    ///
-    ///     POST /api/Accounts/CreateAccount
-    ///     {
-    ///         "ownerId": "00000000-0000-0000-0000-000000000000",
-    ///         "type": 1,
-    ///         "currency": "USD",
-    ///         "interestRate": 3.5
-    ///     }
-    /// </remarks>
+    /// <param name="createAccountDto">Данные для создания счета.</param>
+    /// <returns>Возвращает созданный счет с кодом 201 Created, внутри MbResult.</returns>
     [HttpPost]
-    [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(AccountDto))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(StatusCodes.Status409Conflict)]
+    [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(MbResult<AccountDto>))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(MbResult<string>))]
+    [ProducesResponseType(StatusCodes.Status409Conflict, Type = typeof(MbResult<string>))]
     public async Task<IActionResult> CreateAccount([FromBody] CreateAccountDto createAccountDto)
     {
         var response = await mediator.Send(new CreateAccountCommand(createAccountDto));
 
-        return StatusCode(StatusCodes.Status201Created, response);
+        return Created(string.Empty, response); // assuming response is MbResult<AccountDto>
     }
 
     /// <summary>
     /// Частично обновить данные счета.
     /// </summary>
     /// <param name="accountId">Идентификатор счета для обновления (Guid).</param>
-    /// <param name="updatedAccountDto">Данные для обновления счета. Поля необязательны:
-    /// <list type="bullet">
-    /// <item><description>InterestRate — новая процентная ставка (опционально).</description></item>
-    /// <item><description>ClosedAt — дата закрытия счета (опционально).</description></item>
-    /// </list>
-    /// </param>
-    /// <returns>Возвращает обновленный счет с кодом 200 OK.</returns>
-    /// <remarks>
-    /// Пример запроса:
-    /// 
-    ///     PATCH /api/Accounts/UpdateAccount/{accountId}
-    ///     {
-    ///         "interestRate": 4.0,
-    ///         "closedAt": "2025-07-27T00:00:00Z"
-    ///     }
-    /// </remarks>
+    /// <param name="updatedAccountDto">Данные для обновления счета.</param>
+    /// <returns>Возвращает обновленный счет с кодом 200 OK, внутри MbResult.</returns>
     [HttpPatch("{accountId:guid}")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(AccountDto))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status409Conflict)]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(MbResult<AccountDto>))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(MbResult<string>))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(MbResult<string>))]
+    [ProducesResponseType(StatusCodes.Status409Conflict, Type = typeof(MbResult<string>))]
     public async Task<IActionResult> UpdateAccount(Guid accountId, [FromBody] UpdateAccountDto updatedAccountDto)
     {
         var response = await mediator.Send(new UpdateAccountCommand(accountId, updatedAccountDto));
@@ -89,10 +58,10 @@ public class AccountsController(IMediator mediator) : ControllerBase
     /// Удалить счет по идентификатору.
     /// </summary>
     /// <param name="accountId">Идентификатор удаляемого счета (Guid).</param>
-    /// <returns>Возвращает код 204 No Content при успешном удалении.</returns>
+    /// <returns>Возвращает код 204 No Content при успешном удалении или ошибку внутри MbResult.</returns>
     [HttpDelete("{accountId:guid}")]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(MbResult<bool>))]
     public async Task<IActionResult> DeleteAccount(Guid accountId)
     {
         await mediator.Send(new DeleteAccountCommand(accountId));
@@ -103,33 +72,28 @@ public class AccountsController(IMediator mediator) : ControllerBase
     /// <summary>
     /// Получить список всех счетов.
     /// </summary>
-    /// <returns>Возвращает список счетов с кодом 200 OK.</returns>
+    /// <returns>Возвращает список счетов с кодом 200 OK, внутри MbResult.</returns>
     [HttpGet]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<AccountDto>))]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(MbResult<IEnumerable<AccountDto>>))]
     public async Task<IActionResult> GetAccounts()
     {
         var response = await mediator.Send(new GetAllAccountsQuery());
 
-        return Ok(response.ToList());
+        return Ok(response);
     }
 
     /// <summary>
     /// Получить выписку по счетам владельца за заданный период.
     /// </summary>
     /// <param name="ownerId">Идентификатор владельца счетов (Guid).</param>
-    /// <param name="accountId">Опциональный идентификатор конкретного счета (Guid?).</param>
-    /// <param name="startDate">Дата начала периода выписки (включительно).</param>
-    /// <param name="endDate">Дата окончания периода выписки (включительно).</param>
-    /// <returns>Возвращает выписку по счетам с кодом 200 OK.</returns>
-    /// <remarks>
-    /// Пример запроса:
-    ///
-    ///     GET /api/Accounts/GetAccountStatement/{ownerId}?accountId={accountId}&amp;startDate=2025-01-01&amp;endDate=2025-06-30
-    /// </remarks>
+    /// <param name="accountId">Опциональный идентификатор счета (Guid?).</param>
+    /// <param name="startDate">Дата начала периода (включительно).</param>
+    /// <param name="endDate">Дата конца периода (включительно).</param>
+    /// <returns>Возвращает выписку по счетам с кодом 200 OK, внутри MbResult.</returns>
     [HttpGet("{ownerId:guid}")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(MultiAccountStatementDto))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(MbResult<MultiAccountStatementDto>))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(MbResult<string>))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(MbResult<string>))]
     public async Task<IActionResult> GetAccountStatement(Guid ownerId, [FromQuery] Guid? accountId,
         DateTime startDate, DateTime endDate)
     {
@@ -142,9 +106,9 @@ public class AccountsController(IMediator mediator) : ControllerBase
     /// Проверить существование счета по идентификатору.
     /// </summary>
     /// <param name="accountId">Идентификатор счета (Guid).</param>
-    /// <returns>Возвращает true, если счет существует, иначе false, с кодом 200 OK.</returns>
+    /// <returns>Возвращает true/false с кодом 200 OK, внутри MbResult.</returns>
     [HttpGet("{accountId:guid}")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(bool))]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(MbResult<bool>))]
     public async Task<IActionResult> AccountExists(Guid accountId)
     {
         var response = await mediator.Send(new CheckAccountExistsQuery(accountId));
