+++ SimpleAccountService/Features/Accounts/Commands/DeleteAccount/DeleteAccountCommandHandler.cs
@@ -1,14 +1,23 @@
 ﻿using JetBrains.Annotations;
 using MediatR;
-using Simple_Account_Service.Features.Accounts.Interfaces;
+using Simple_Account_Service.Application.Exceptions;
+using Simple_Account_Service.Application.Models;
+using Simple_Account_Service.Features.Accounts.Interfaces.Repositories;
 
 namespace Simple_Account_Service.Features.Accounts.Commands.DeleteAccount;
 
 [UsedImplicitly]
-public class DeleteAccountCommandHandler(IAccountsService service) : IRequestHandler<DeleteAccountCommand>
+public class DeleteAccountCommandHandler(IAccountRepository repository) : IRequestHandler<DeleteAccountCommand, MbResult<bool>>
 {
-    public async Task Handle(DeleteAccountCommand request, CancellationToken cancellationToken)
+    public async Task<MbResult<bool>> Handle(DeleteAccountCommand request, CancellationToken cancellationToken)
     {
-        await service.DeleteAccountAsync(request.AccountId);
+        var account = await repository.GetByIdAsync(request.AccountId);
+
+        if (account == null)
+        {
+            throw new NotFoundException($"Счет с айди {request.AccountId} не найден.");
+        }
+
+        return new MbResult<bool>(await repository.DeleteAsync(request.AccountId));
     }
 }
