+++ SimpleAccountService/Features/Accounts/Commands/UpdateAccount/UpdateAccountCommand.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
+using Simple_Account_Service.Application.Models;
 
 namespace Simple_Account_Service.Features.Accounts.Commands.UpdateAccount;
 
-public record UpdateAccountCommand(Guid AccountId, UpdateAccountDto Request) : IRequest<AccountDto>;
+public record UpdateAccountCommand(Guid AccountId, UpdateAccountDto Request) : IRequest<MbResult<AccountDto>>;
