+++ SimpleAccountService/Features/Accounts/Queries/GetAccountStatement/GetAccountStatementQuery.cs
@@ -1,7 +1,8 @@
 ﻿using MediatR;
+using Simple_Account_Service.Application.Models;
 using Simple_Account_Service.Features.Accounts.Queries.GetAccountStatement.Dto;
 
 namespace Simple_Account_Service.Features.Accounts.Queries.GetAccountStatement;
 
 public record GetAccountStatementQuery(Guid OwnerId, Guid? AccountId,
-    DateTime StartDate, DateTime EndDate) : IRequest<MultiAccountStatementDto>;
+    DateTime StartDate, DateTime EndDate) : IRequest<MbResult<MultiAccountStatementDto>>;
