+++ SimpleAccountService/Features/Accounts/Queries/GetAccounts/GetAllAccountsQuery.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
+using Simple_Account_Service.Application.Models;
 
 namespace Simple_Account_Service.Features.Accounts.Queries.GetAccounts;
 
-public record GetAllAccountsQuery : IRequest<IEnumerable<AccountDto>>;
+public record GetAllAccountsQuery : IRequest<MbResult<IEnumerable<AccountDto>>>;
