+++ launchSettings.json
@@ -1,4 +1,4 @@
-{
+﻿{
   "profiles": {
     "Docker Compose": {
       "commandName": "DockerCompose",
